feat: Implement client-server communication protocol
The code changes include implementing the communication protocol between the client and server. This involves sending the initial `<hello>` message from the server and waiting for the correct response from the client. If the response is correct, the server sends the `<login> <ack>` command to confirm the connection.

Recent user commits:
- Add tmp folder creation in win-build.bat
- Preparing for client-server implementation
- Update input method names
- Corrigido tratamento de entrada do usuário e exibição do menu
- Corrigido tratamento de entrada do usuário e exibição do menu
- Refactored code
- Working arrow keys under Windows
- Working arrow keys under Unix
- Refactored setConsoleMode method in OS.java fto ensure Unix compatibility
- Refactored menu selection logic in Main.java and drawBox method in Output.java